services/ai_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
pydantic==2.10.4
httpx==0.28.1
boto3==1.34.0
langchain>=0.1.16
langchain-google-genai>=1.0.3
langchain-groq>=0.1.3
langchain-chroma>=0.1.0
chromadb==0.5.0
